Refine agent-context adapter boundaries
Move JSON rendering and downstream CLI literals out of the core
`AgentContext` model. Split bounded test modules, strengthen output
oracles, stabilize the bridge integration test, and align the associated
documentation and ExecPlan with the public surface.
