+++ cargo-orthohelp/tests/nested_subcommand_end_to_end.rs
@@ -16,12 +16,17 @@ use std::error::Error;
 use std::fmt::{self, Display, Formatter};
 use std::io::Read;
 use std::process::Command;
+use std::sync::{LazyLock, Mutex};
 use tempfile::TempDir;
 
 type TestError = Box<dyn Error + Send + Sync>;
 
 const FIXTURE_PACKAGE: &str = "orthohelp_fixture";
 const FIXTURE_ROOT_TYPE: &str = "orthohelp_fixture::NestedFixtureConfig";
+
+// The rstest cases share a content-addressed bridge directory whose manifest is
+// rewritten before each build, so their subprocesses must not overlap.
+static BRIDGE_BUILD_MUTEX: LazyLock<Mutex<()>> = LazyLock::new(Mutex::default);
 
 #[rstest]
 #[case::ir(
@@ -57,6 +62,9 @@ fn nested_subcommand_tree_survives_bridge_outputs(
     #[case] args: &[&str],
     #[case] assertion: fn(&Utf8PathBuf) -> Result<(), TestError>,
 ) -> Result<(), TestError> {
+    let _bridge_build_guard = BRIDGE_BUILD_MUTEX
+        .lock()
+        .expect("bridge build test mutex is not poisoned");
     let (_temp, out_dir) = temp_out_dir()?;
     run_orthohelp(&out_dir, args)?;
     assertion(&out_dir).map_err(|err| assertion_failed(name, err))?;
